Move 'create_db' out of app_factory.py and into orm.py
